requirements.txt
Pillow==9.1.0
seaborn==0.11.2
Sphinx==4.5.0
pytest==7.1.2
pytesseract==0.3.9
numpy==1.22.3
setuptools==62.1.0
matplotlib==3.5.2
python-dotenv==0.20.0
pathlib==1.0.1